requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.37.37
botocore==1.37.37
cffi==1.17.1
channels==4.0.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.0.0
Django==5.0.13
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
mongoengine==0.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pymongo==4.11.2
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
s3transfer==0.11.5
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
